infrastructure/developmentServer/scripts/requirements.txt
hcloud==1.22.0
python-dotenv==1.0.0
